pipeline/requirements.txt
requests==2.31.0
vaderSentiment==3.3.2
supabase==2.15.0
python-dotenv==1.0.1
